Thumb: Move implementations of operations under thumb module
