+++ emu/src/cpu/thumb/mod.rs
@@ -1,3 +1,4 @@
 pub mod alu_instructions;
 pub mod instruction;
 pub mod mode;
+pub mod operations;
